An initial retrieval of approved applicants over the last N days.
This is really just a proof of concept.  We'll eventually
be adding approval dates to previously existing data regarding
applicants.

To run this, the container must be started:

  docker compose --ansi never up --build

  (or in detached mode with -d)

Run as:

  docker exec -it background-check-app bin/testGetForDate.0.py --days 10 --district dd --county ccc

which gets the last ten days of approved applicants.

The returned data is in JSON, but this is just a demonstration.
What happens next with the data is TBD.

Another possible change is replacing '--days' (the number of
days back to check) with a start and end date.  Which is better
depends upon how we're using this (or the utility could support
both).